minimal changes to render genome-aligned heatmaps
This track type was motivated by _multiscale_ genomic analyses wherein the
same reduction is applied in multiple "window" sizes that are swept across
a genome with the goal of capturing the size most relevant to a phenomena
of study.

A heatmap typically maps continuous data to one of two types of color
palettes, depending on the distribution of the data to be visualized:
1. a "sequential" palette, dark to light or mute to saturated, for data
	from a one-sided distributions (e.g. count data)
2. a "diverging" palette with contrasting colors at the upper and lower
	extremes and a mute color in the center for data from two-sided
	distributions (e.g. Gaussian)

The submitted pull request tries to balance:
1. minimizing invasive changes to existing IGV code
2. minimizing impact on performance
3. maximizing generality of the new capability (add capability, not policy)
4. minimizing new requirements on the user

Because a heatmap can be thought of as nothing more than multiple rows of
densely-packed annotation features (typically _without_ any additional
decoration like labels), the FeatureTrack that supports general annotations
already has all facilities required to support heatmaps. In particular,
1. All the behavior concerning feature _visibility_ is used as-is.
2. Its ability to render multiple _rows_ within one track is used.
3. Only some small modifications to its behavior were necessary.
	* Input data, not packFeatures, _assigns_ Features to rows.
	* No decoration of heatmap cells is supported.
	* Data is mapped at runtime to colors used to fill heatmap cells.

These changes have almost no impact on performance, and what little
performance impact there is could be mitigated with slightly more invasive
changes.

To keep the IGV implementation as simple (and fast) as possible, the data
is expected to be fully preprocessed for display; all IGV does at runtime
is map numeric values to colors using a colormap function and a palette.
Concretely, data should be in [0,1]. Values below and above this range are
by default clamped to 0 and 1 respectively and thus mapped to the palette's
edge colors. Alternatively, 2 discrete "outlier" colors can be provided in
the track config to highlight outlying data.

Given the preceding characterization of heatmaps, it is natural to
deliver heatmap data as BED files with a very minor abuse of the format:
the 4th (name) column contains a 0-based row _assignment_.
The name BED field can be thought of as naming the row. Since genome
coordinate ranges in heatmap data would not typically be associated with
other names, this is not such an abuse of the BED format.
The 5th (score) columns is used for it's intended purpose: a score.

This arrangement also allows additional runtime optimizations:
1. the mute color corresponding to the most "uninteresting" data range
	is provided by the config.altColor. Thus, no cells that would be mapped
	to this value need to be included in the input data! The data can be
	minified.
2. Obviously, the _range_ of values mapped to this color can be adjusted
	at data preparation time to effectively compress the data with minimal
	loss of information (similar what is done in preparation of a JPG image).
3. Similarly, adjacent cells that are not "too" different could be
	optionally merged during data preparation to further reduce size.

Obviously all these optimizations are entirely optional.

Only one new JavaScript file, heatmap.js, is added providing:
1. a default, linear colormap function and
2. a small selection of 64-color palettes (some adapted from Matplotlib and
	others generated by the colorspace package from R)
3. the renderCell function called by FeatureTrack.draw.

Only the renderCell function is necessary. The colormap function and
palettes could be made the user's responsibility to be defined in the
config, but as a suitable palette and colormap function is always necessary
and a linear map is most common, providing these as defaults reduces work
for user. Importantly _both can still be defined entirely in the config,
maximizing generality_.

With the above considerations only a few edits to IGV were necessary:

1. A new config type is added: "heatmap"
2. Exports from heatmap.js are exposed in index.js to make the default
	colormap and palettes available to the user in their config.
3. A one-line change to trackFactory.js mapping "heatmap" to "feature".
4. A conditional in TextFeatureSource.loadFeatures that parses the 4th
	column of a BED file as an integer and assigns the resulting number to
	the (already-existing) row attribute of Feature.
5. Only a few significant changes in FeatureTrack:
	1. setting renderCell as the FeatureTrack.render method
	2. setting background color
	3. preclude the code that "Ensure[s] a visible gap between features"

The following _should_ be set in the track config:
1. maxRows - the number of rows in the heatmap
2. height - the pixel height of the heatmap track. This is used verbatim.
	In particular, displayMode and its related variables are _not_ used by
	heatmap tracks.
3. color - a function taking a Feature as it's sole argument.
4. altColor - used as the "background" and support sparser input data. This
	color should typically correspond to the most "uninteresting" color in the
	heatmap's palette.

Defaults are provided for everything that insure _something_ is displayed,
though it will certainly not be ideal without user configuration, and it
won't even be _correct_ if maxRows is unset.

The maxRows config element could be made optional since largest row index can
be determined at runtime, but requiring specification of maxRows simplifies the
implementation and is not unreasonable given that dimensions are always part of
a heatmap's definition. In this case, one dimension is implicity genomic; the
other defined by the depth of user's data.

As is, the implementation simply make full use of configured space, so
heatmap lines are config.height / config.maxRows pixels. In particular,
squishedRowHeight and related config variables are ignored, and no runtime
adjustment of track height should occur.

It may be undesirable to expose the Feature class to the config.color
function, but this maximizes generality. Users could accommodate non-unitary
scores (outside [0,1]) or even consider strand in the color mapping. Also,
by essentially requiring the user's function to pull the score member out of
Feature, we decouple Feature from the colormap function which _really_
should only concern itself with a number and colors.

Input must come from BED files with:
1. 0-based row number as the _sole content_ of the 4th column, and
2. score in [0,1] in the 5th column
3. The 6th (strand) column is ignored

The content of the 4th column certainly could be generalized to, for example,
accommodate a "row=N" string as one of a set of semi-colon delimited args
(since GFF3 column 9-style data is already supported in BED column 4).